requirements.txt
# Google ADK and Vertex AI
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.3.0
vertexai>=1.40.0

# Google Workspace APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Airtable
pyairtable>=2.1.0

# Cloud Services
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
tenacity>=8.2.3
pyyaml>=6.0.1

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0
reportlab>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Type hints
typing-extensions>=4.8.0

